requirements.txt
pandas==2.1.3
streamlit==1.28.2
paho-mqtt==1.6.1
pyserial==3.5
ephem==4.1.5
pysqlite3

